organize routes a lil' bit
