+++ .gitignore
@@ -6,4 +6,4 @@ build
 *.out.*
 artifacts
 .local
-api
+/api
