+++ app/modules/media/media-card.tsx
@@ -4,7 +4,7 @@ import { filterJoin } from "~/modules/common/filter-join"
 import { infix } from "~/modules/common/infix"
 import { MediaCardControls } from "~/modules/media/media-card-controls"
 import { LazyImage } from "~/modules/ui/lazy-image"
-import { getOptimizedImageUrl } from "~/routes/optimized-image"
+import { getOptimizedImageUrl } from "~/routes/api/optimized-image"
 import type { AnilistMedia } from "./media-data"
 
 export const MediaCard = memo(function MediaCard({
